EditDetailsForm: rename from EditConnectionForm
