--- src/components/family-connections/EditConnectionForm/index.js
@@ -1,2 +1,0 @@
-import EditConnectionForm from './EditConnectionForm';
-export default EditConnectionForm;
