+++ src/components/family-connections/EditDetailsForm/EditDefailsForm.test.js
@@ -4,19 +4,19 @@ import { Platform, View } from 'react-native';
 import renderer from 'react-test-renderer';
 import { render, fireEvent } from 'react-native-testing-library';
 
-import EditConnectionForm from './EditConnectionForm';
+import EditDetailsForm from './EditDetailsForm';
 
 /**********************************************************/
 
-describe('EditConnectionForm component', () => {
+describe('EditDetailsForm component', () => {
 
     test('renders', () => {
-        const tree = renderer.create(<EditConnectionForm/>).toJSON();
+        const tree = renderer.create(<EditDetailsForm/>).toJSON();
         expect(tree.children).toHaveLength(1);
     });
 
     test('matches snapshot', () => {
-        const tree = renderer.create(<EditConnectionForm/>).toJSON();
+        const tree = renderer.create(<EditDetailsForm/>).toJSON();
         expect(tree).toMatchSnapshot();
     });
