+++ src/components/family-connections/EditDetailsForm/index.js
@@ -1,0 +1,2 @@
+import EditDetailsForm from './EditDetailsForm';
+export default EditDetailsForm;
